requirements.txt
flask==2.3.2
werkzeug==2.3.6
torch==2.0.0
torchvision==0.15.1
pillow==9.5.0
numpy==1.24.3
efficientnet-pytorch==0.7.1
gunicorn==21.2.0
flask-cors==4.0.0
